requirements.txt
fastapi
pydantic
uvicorn
requests
python-dotenv
pandas
scikit-learn
numpy
openrouteservice

